Deprecated functions in package ‘derfinder’

These functions are provided for compatibility with older version of ‘derfinder’ only and will be defunct at the next release.

advancedArg(fun, package = "derfinder", browse = interactive())

Arguments

fun

The name of a function(s) that has advanced arguments in package.

package

The name of the package where the function is stored. Only 'derfinder', 'derfinderPlot', and 'regionReport' are accepted.

browse

Whether to open the URLs in a browser.

Value

A vector of URLs with the GitHub search queries.

Details

The following functions are deprecated and will be made defunct.

advancedArg Not needed now that the advanced arguments are documented on the help pages of each function.

Examples

## Open the advanced argument docs for loadCoverage()
if (interactive()) {
    advancedArg("loadCoverage")
} else {
    (advancedArg("loadCoverage", browse = FALSE))
}
#> Warning: This function is not needed anymore now that the advanced arguments are document in the help pages of each function.
#> [1] "https://github.com/search?utf8=%E2%9C%93&q=advanced_argument+filename%3Aload+repo%3Alcolladotor%2Fderfinder+path%3A%2FR&type=Code"